requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pyyaml==6.0.2
openai==1.50.0
requests==2.32.0
pytest==8.3.0
